// src/types/ApprovalResponse.ts
// Generated by https://quicktype.io

export interface PRActivityResponse {
    values: Value[];
    pagelen: number;
}

export interface Value {
    pull_request: Pullrequest;
    update?: Update;
    comment?: Comment;
    approval?: Approval;
}

export interface Approval {
    date: string;
    user: User;
    pullrequest: Pullrequest;
}

export interface Pullrequest {
    type: string;
    id: number;
    title: string;
    links: PullRequestLinks;
}

export interface PullRequestLinks {
    self: HTML;
    html: HTML;
}

export interface HTML {
    href: string;
}

export interface User {
    display_name: string;
    links: UserLinks;
    type: string;
    uuid: string;
    account_id: string;
    nickname: string;
}

export interface UserLinks {
    self: HTML;
    avatar: HTML;
    html: HTML;
}

export interface Comment {
    id: number;
    created_on: string;
    updated_on: string;
    content: Content;
    user: User;
    deleted: boolean;
    parent?: Parent;
    inline: string;
    type: string;
    links: PullRequestLinks;
    pullrequest: Pullrequest;
}

export interface Content {
    type: string;
    raw: string;
    markup: string;
    html: string;
}

export interface Parent {
    id: number;
    links: PullRequestLinks;
}

export interface Update {
    state: string;
    title: string;
    description: string;
    reviewers: any[];
    changes: Changes;
    reason: string;
    author: User;
    date: string;
    destination: Destination;
    source: Destination;
}

export interface Changes {
    status?: Status;
}

export interface Status {
    old: string;
    new: string;
}

export interface Destination {
    branch: Branch;
    commit: Commit;
    repository: Repository;
}

export interface Branch {
    name: string;
}

export interface Commit {
    type: string;
    hash: string;
    links: PullRequestLinks;
}

export interface Repository {
    type: string;
    full_name: string;
    links: UserLinks;
    name: string;
    uuid: string;
}